requirements.txt
flask==2.0.1
pandas==1.3.5
numpy==1.21.6
yfinance==0.1.70
scikit-learn==1.0.2
nltk==3.8.1
gunicorn==20.1.0
requests==2.31.0
newsapi-python==0.2.7
textblob==0.17.1
flask-cors==4.0.0 